[remote-bindings] Relocate preview shim so the move renders as a rename
Delete wrangler's create-worker-preview.ts and move the thin logger/auth
injecting wrappers to dev/preview.ts, repointing importers. Freeing the old
path lets git pair it with the package's create-worker-preview.ts (82%
similar), so GitHub renders the extraction as a rename instead of a
delete + add.
