+++ packages/wrangler/src/api/startDevWorker/RemoteRuntimeController.ts
@@ -7,7 +7,7 @@ import { version as packageVersion } from "../../../package.json";
 import {
 	createPreviewSession,
 	createWorkerPreview,
-} from "../../dev/create-worker-preview";
+} from "../../dev/preview";
 import {
 	createRemoteWorkerInit,
 	getWorkerAccountAndContext,
@@ -26,7 +26,7 @@ import type {
 	CfAccount,
 	CfPreviewSession,
 	CfPreviewToken,
-} from "../../dev/create-worker-preview";
+} from "../../dev/preview";
 import type {
 	BundleCompleteEvent,
 	BundleStartEvent,
